src/skills/renaming-file-personal-style/requirements.txt
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.0.0
chandra-ocr[hf]>=2.0.0,<3.0.0
